Time: 29 ms (43.96%), Space: 99.3 MB (27.25%) - LeetHub

+++ 1590-make-sum-divisible-by-p/1590-make-sum-divisible-by-p.java
@@ -1,24 +1,24 @@
 class Solution {
     public int minSubarray(int[] nums, int p) {
-        int sum=0;
+        long sum=0;
         for(int i=0;i<nums.length;i++){
             sum+=nums[i];
         }
-        int need=sum%p;
+        int need=(int)(sum%p);
         if(need==0){
             return 0;
         }
         HashMap<Integer , Integer> map=new HashMap<>();
         map.put(0,-1);
-        int prefix=0;
+        long prefix=0;
         int ans=nums.length;
         for(int i=0;i<nums.length;i++){
             prefix=(prefix+nums[i])%p;
-            int target=(prefix-need+p)%p;
+            int target=(int)((prefix-need+p)%p);
             if(map.containsKey(target)){
                 ans=Math.min(ans,i-map.get(target));
             }
-            map.put(prefix,i);
+            map.put((int)prefix,i);
         }
         return ans==nums.length ? -1 :ans;
